+++ src/backend/clip/clipManager.js
@@ -1,0 +1,37 @@
+const fs = require("fs");
+const path = require("path");
+
+class ClipManager {
+
+    constructor() {
+        this.projectJsonPath = null;
+        this.project = null;
+    }
+
+    load(projectJsonPath) {
+
+    }
+
+    save() {
+
+    }
+
+    addClip(clip) {
+
+    }
+
+    removeClip(id) {
+
+    }
+
+    renameClip(id, newName) {
+
+    }
+
+    getClips() {
+        return [];
+    }
+
+}
+
+module.exports = new ClipManager();